Will move 'RemoteFeedItem' to its own file
