+++ ModularFeedTDDCaseStudy/ModularFeedTDDCaseStudy/FeedAPI/FeedItemsMapper.swift
@@ -8,13 +8,6 @@
 
 import Foundation
 
-internal struct RemoteFeedItem: Decodable {
-    internal let id: UUID
-    internal let description: String?
-    internal let location: String?
-    internal let image: URL
-}
-
 internal final class FeedItemsMapper {
     
     // This item is the internal representation of a FeedItem, but for the api module.
