+++ ModularFeedTDDCaseStudy/ModularFeedTDDCaseStudy/FeedAPI/RemoteFeedItem.swift
@@ -1,0 +1,16 @@
+//
+//  RemoteFeedItem.swift
+//  ModularFeedTDDCaseStudy
+//
+//  Created by Clint M on 3/22/19.
+//  Copyright © 2019 Clint M. All rights reserved.
+//
+
+import Foundation
+
+internal struct RemoteFeedItem: Decodable {
+    internal let id: UUID
+    internal let description: String?
+    internal let location: String?
+    internal let image: URL
+}
